Series: fix View Series 500 (fractional positions) + speed up; hide metadata sources in UI
- View Series returned nothing because SeriesBook.position was typed int but
  Hardcover series carry fractional positions for novellas (2.5, 9.5, 17.6) →
  Pydantic 500. Position is now float. (This was the "nothing returned" bug.)
- detect_series is much faster: lead with the fast authoritative Hardcover
  book_series lookup (skips the slow OL series-name confirmation when it hits),
  run the OL/GB supplements concurrently under a 4s timeout, and cache the
  enumeration per title (DB hooked/catalog status re-annotated fresh) so repeat
  "View Series" clicks are instant. (~6s → sub-second on cache hit.)
- Metadata sources (Google Books / Open Library / Hardcover) are now hidden from
  the catalog card AND detail modal — they're backend-only listings, never shown
  as selectable sources. Acquire + View Series use the group's representative id
  so they work even when every source is a (hidden) listing.

Tests: fractional series position serializes via SeriesOut; series enumeration
ordering incl. novella; listing sources filtered. 436 passed; frontend builds.

+++ backend/app/schemas.py
@@ -489,7 +489,7 @@ class SeriesBookOut(BaseModel):
     title: str
     author: str | None = None
     year: int | None = None
-    position: int | None = None
+    position: float | None = None   # fractional positions exist for novellas (e.g. 2.5, 12.5)
     cover_url: str | None = None
     ref: str | None = None              # Open Library work key (stable selector)
     catalog_id: int | None = None
